requirements.txt
streamlit==1.29.0
pandas==2.1.4
cryptography==41.0.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
